cloud_run_data_vendor/scripts/requirements.txt
firebase_admin
sendgrid
google-cloud-secret-manager
nylas